+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+
+## [1.0.2] 
+- Update storage to use hexString instead of a string of List<int>
+- Made sure moving from GetStorage to/from GetSecureStorage works and recover from wrong pwd
+
 ## [1.0.0] 
 - Initial release and port of v2.1.0 of GetStorage